refactor: move other controller commands into folders by similarity
