+++ packages/zwave-js/src/lib/commandclass/Security2CC.ts
@@ -24,11 +24,11 @@ import {
 } from "@zwave-js/core";
 import { buffer2hex, getEnumMemberName, pick } from "@zwave-js/shared";
 import type { ZWaveController } from "../controller/Controller";
-import { SendDataBridgeRequest } from "../controller/SendDataBridgeMessages";
-import { SendDataRequest } from "../controller/SendDataMessages";
 import { TransmitOptions } from "../controller/_Types";
 import type { Driver } from "../driver/Driver";
 import { FunctionType, MessagePriority } from "../message/Constants";
+import { SendDataBridgeRequest } from "../serialapi/transport/SendDataBridgeMessages";
+import { SendDataRequest } from "../serialapi/transport/SendDataMessages";
 import { CCAPI } from "./API";
 import {
 	API,
