+++ packages/zwave-js/src/lib/commandclass/ZWavePlusCC.test.ts
@@ -1,8 +1,8 @@
 import { CommandClasses } from "@zwave-js/core";
-import { SendDataRequest } from "../controller/SendDataMessages";
 import { TransmitOptions } from "../controller/_Types";
 import type { Driver } from "../driver/Driver";
 import { ZWaveNode } from "../node/Node";
+import { SendDataRequest } from "../serialapi/transport/SendDataMessages";
 import { assertCC } from "../test/assertCC";
 import { createEmptyMockDriver } from "../test/mocks";
 import { CommandClass, getCommandClass } from "./CommandClass";
