+++ packages/zwave-js/src/lib/driver/CommandQueueMachine.test.ts
@@ -3,9 +3,12 @@ import { SecurityManager } from "@zwave-js/core";
 import { createDeferredPromise } from "alcalzone-shared/deferred-promise";
 import { assign, interpret, Machine, State } from "xstate";
 import { BasicCCGet, BasicCCReport, BasicCCSet } from "../commandclass/BasicCC";
-import { SendDataAbort, SendDataRequest } from "../controller/SendDataMessages";
 import { MessagePriority } from "../message/Constants";
 import type { Message } from "../message/Message";
+import {
+	SendDataAbort,
+	SendDataRequest,
+} from "../serialapi/transport/SendDataMessages";
 import { dummyCallbackNOK, dummyResponseNOK } from "../test/messages";
 import { createEmptyMockDriver } from "../test/mocks";
 import {
