+++ packages/zwave-js/src/lib/driver/MessageGenerators.ts
@@ -15,8 +15,11 @@ import {
 	SecurityCCNonceGet,
 	SecurityCCNonceReport,
 } from "../commandclass/SecurityCC";
-import { isSendData, isTransmitReport } from "../controller/SendDataShared";
 import type { Message } from "../message/Message";
+import {
+	isSendData,
+	isTransmitReport,
+} from "../serialapi/transport/SendDataShared";
 import type { Driver, SendCommandOptions } from "./Driver";
 import { sendDataErrorToZWaveError } from "./StateMachineShared";
 import type { MessageGenerator } from "./Transaction";
