+++ packages/zwave-js/src/lib/driver/Transaction.test.ts
@@ -1,11 +1,11 @@
 import { NoOperationCC } from "../commandclass/NoOperationCC";
-import { GetControllerVersionRequest } from "../controller/GetControllerVersionMessages";
-import { SendDataRequest } from "../controller/SendDataMessages";
 import { MessagePriority } from "../message/Constants";
 import { getDefaultPriority, Message } from "../message/Message";
 import type { ZWaveNode } from "../node/Node";
 import { NodeStatus } from "../node/_Types";
+import { GetControllerVersionRequest } from "../serialapi/capability/GetControllerVersionMessages";
 import { RemoveFailedNodeRequest } from "../serialapi/network-mgmt/RemoveFailedNodeMessages";
+import { SendDataRequest } from "../serialapi/transport/SendDataMessages";
 import type { Driver } from "./Driver";
 import {
 	MessageGenerator,
