+++ packages/zwave-js/src/lib/driver/TransactionMachine.ts
@@ -9,14 +9,14 @@ import {
 	StateMachine,
 } from "xstate";
 import { messageIsPing } from "../commandclass/NoOperationCC";
-import { SendDataMulticastBridgeRequest } from "../controller/SendDataBridgeMessages";
-import { SendDataMulticastRequest } from "../controller/SendDataMessages";
+import type { Message } from "../message/Message";
+import { SendDataMulticastBridgeRequest } from "../serialapi/transport/SendDataBridgeMessages";
+import { SendDataMulticastRequest } from "../serialapi/transport/SendDataMessages";
 import {
 	isSendData,
 	isTransmitReport,
 	SendDataMessage,
-} from "../controller/SendDataShared";
-import type { Message } from "../message/Message";
+} from "../serialapi/transport/SendDataShared";
 import type { CommandQueueEvent } from "./CommandQueueMachine";
 import {
 	createMessageDroppedUnexpectedError,
