+++ packages/zwave-js/src/lib/node/Node.ts
@@ -150,11 +150,6 @@ import {
 	ZWavePlusNodeType,
 	ZWavePlusRoleType,
 } from "../commandclass/_Types";
-import {
-	ApplicationUpdateRequest,
-	ApplicationUpdateRequestNodeInfoReceived,
-	ApplicationUpdateRequestNodeInfoRequestFailed,
-} from "../controller/ApplicationUpdateRequest";
 import { isRssiError, RSSI, RssiError, TXReport } from "../controller/_Types";
 import type { Driver, SendCommandOptions } from "../driver/Driver";
 import { cacheKeys } from "../driver/NetworkCache";
@@ -163,6 +158,11 @@ import type { StatisticsEventCallbacksWithSelf } from "../driver/Statistics";
 import type { Transaction } from "../driver/Transaction";
 import { MessagePriority } from "../message/Constants";
 import type { Message } from "../message/Message";
+import {
+	ApplicationUpdateRequest,
+	ApplicationUpdateRequestNodeInfoReceived,
+	ApplicationUpdateRequestNodeInfoRequestFailed,
+} from "../serialapi/application/ApplicationUpdateRequest";
 import {
 	GetNodeProtocolInfoRequest,
 	type GetNodeProtocolInfoResponse,
