+++ packages/zwave-js/src/lib/serialapi/application/BridgeApplicationCommandRequest.test.ts
@@ -1,8 +1,8 @@
 import { ConfigManager } from "@zwave-js/config";
-import "../commandclass/index";
-import type { Driver } from "../driver/Driver";
-import { Message } from "../message/Message";
-import { createEmptyMockDriver } from "../test/mocks";
+import "../../commandclass/index";
+import type { Driver } from "../../driver/Driver";
+import { Message } from "../../message/Message";
+import { createEmptyMockDriver } from "../../test/mocks";
 
 const fakeDriver = createEmptyMockDriver() as unknown as Driver;
