+++ packages/zwave-js/src/lib/serialapi/application/BridgeApplicationCommandRequest.ts
@@ -4,23 +4,23 @@ import {
 	NODE_ID_BROADCAST,
 } from "@zwave-js/core";
 import { getEnumMemberName } from "@zwave-js/shared";
-import { CommandClass, SinglecastCC } from "../commandclass/CommandClass";
-import type { ICommandClassContainer } from "../commandclass/ICommandClassContainer";
-import type { Driver } from "../driver/Driver";
+import { CommandClass, SinglecastCC } from "../../commandclass/CommandClass";
+import type { ICommandClassContainer } from "../../commandclass/ICommandClassContainer";
+import { RSSI, RssiError } from "../../controller/_Types";
+import type { Driver } from "../../driver/Driver";
 import {
 	FunctionType,
 	MessagePriority,
 	MessageType,
-} from "../message/Constants";
+} from "../../message/Constants";
 import {
 	Message,
 	MessageDeserializationOptions,
 	messageTypes,
 	priority,
-} from "../message/Message";
+} from "../../message/Message";
+import { tryParseRSSI } from "../transport/SendDataShared";
 import { ApplicationCommandStatusFlags } from "./ApplicationCommandRequest";
-import { tryParseRSSI } from "./SendDataShared";
-import { RSSI, RssiError } from "./_Types";
 
 @messageTypes(MessageType.Request, FunctionType.BridgeApplicationCommand)
 // This does not expect a response. The controller sends us this when a node sends a command
