+++ packages/zwave-js/src/lib/serialapi/capability/SerialAPISetupMessages.test.ts
