+++ packages/zwave-js/src/lib/serialapi/capability/SerialAPISetupMessages.ts
