+++ packages/zwave-js/src/lib/serialapi/misc/SoftResetRequest.ts
@@ -2,8 +2,8 @@ import {
 	FunctionType,
 	MessagePriority,
 	MessageType,
-} from "../message/Constants";
-import { Message, messageTypes, priority } from "../message/Message";
+} from "../../message/Constants";
+import { Message, messageTypes, priority } from "../../message/Message";
 
 @messageTypes(MessageType.Request, FunctionType.SoftReset)
 @priority(MessagePriority.Controller)
